+++ lcp/readium-lcp-swift/LCPAuthenticating.swift
@@ -1,12 +1,7 @@
 //
-//  LCPAuthenticating.swift
-//  r2-lcp-swift
-//
-//  Created by Mickaël Menu on 08.02.19.
-//
-//  Copyright 2019 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Copyright 2020 Readium Foundation. All rights reserved.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation