+++ lcp/CHANGELOG.md
@@ -8,14 +8,14 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-* LCP implementation of the [Content Protection API](https://github.com/readium/architecture/blob/master/proposals/006-content-protection.md) to work with the new [Streamer API](https://github.com/readium/architecture/blob/master/proposals/005-streamer-api.md).
+* LCP implementation of the [Content Protection API](https://readium.org/architecture/proposals/006-content-protection) to work with the new [Streamer API](https://readium.org/architecture/proposals/005-streamer-api).
   * It is highly recommended that you upgrade to the new `Streamer` API to open publications, which will simplify DRM unlocking.
 * Two default implementations of `LCPAuthenticating`:
   * `LCPDialogAuthentication` to prompt the user for its passphrase with the official LCP dialog.
   * `LCPPassphraseAuthentication` to provide directly a passphrase, pulled for example from a database or a web service.
 * `LCPService.acquirePublication()` is a new API to acquire a publication from a standalone license. Compared to the former `importPublication()`:
   * It doesn't require the passphrase, to allow bulk imports.
-  * It can be cancelled by calling `cancel()` on the returned `LCPAuthentication` object.
+  * It can be cancelled by calling `cancel()` on the returned `LCPAcquisition` object.
 * `LCPService.isLCPProtected()` provides a way to check if a file is protected with LCP.
 * `LCPService.addPassphrase()` can be used to preload LCP passphrases, for example when using [LCP Automatic Key Retrieval](https://readium.org/lcp-specs/notes/lcp-key-retrieval.html).
 
@@ -23,6 +23,9 @@ All notable changes to this project will be documented in this file.
 
 * `LCPAuthenticating` is now provided with more information and you will need to update your implementation.
 
+### Fixed
+
+* [Decrypting resources in some edge cases](https://github.com/readium/r2-lcp-swift/pull/94).
 
 ## [2.0.0-alpha.1]