+++ lcp/CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+
+## [2.0.0-alpha.2]
+
 ### Added
 
 * LCP implementation of the [Content Protection API](https://readium.org/architecture/proposals/006-content-protection) to work with the new [Streamer API](https://readium.org/architecture/proposals/005-streamer-api).
@@ -26,6 +29,7 @@ All notable changes to this project will be documented in this file.
 
 * [Decrypting resources in some edge cases](https://github.com/readium/r2-lcp-swift/pull/94).
 
+
 ## [2.0.0-alpha.1]
 
 ### Added
@@ -40,3 +44,4 @@ All notable changes to this project will be documented in this file.
 
 [unreleased]: https://github.com/readium/r2-lcp-swift/compare/master...HEAD
 [2.0.0-alpha.1]: https://github.com/readium/r2-lcp-swift/compare/1.2.3...2.0.0-alpha.1
+[2.0.0-alpha.2]: https://github.com/readium/r2-lcp-swift/compare/2.0.0-alpha.1...2.0.0-alpha.2